+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+Nothing yet.
+
+## [1.2.0] - 2026-09-05
+
 ### Added
 
 - **`mac_required` config key** — makes MAC registration itself